mismatched getter/setter

+++ assay/api-src/org/labkey/api/assay/nab/NabSpecimen.java
@@ -186,7 +186,7 @@ public class NabSpecimen
         return _rowId;
     }
 
-    public void setRowId(int rowId)
+    public void setRowId(long rowId)
     {
         _rowId = rowId;
     }
